Week1.cpp
Week 1 (Review of functions, structures)
{Aim:- Defining functions, invoking functions, passing parameters to
functions and returning values from functions. (all work on integers)}
Write functions for the following and show their invocation in main.
(1) A function to search for an integer in a list of integers. If found the return value of the
function is 1 else return value is -1. Pass the list of integers, length of the array and the key to be
searched as parameters to the function.
(2) A function which determines the biggest element in an array of integers. The function returns
this number. Pass the list of integers and length of the array as parameters to the function.
(3) Write a function which sorts an array of integers in ascending order using bubble sort. The
length of the array and the array are passed as parameters to the function.
(4) Define a structure called Rectangle which keeps track of the length and breadth of a
rectangle. Write functions namely input, displayDimensions, displayArea and edit to input the
dimensions of a rectangle, to display the dimensions, to calculate and display the area of a
triangle respectively. Write a main function which defines a variable of type Rectangle. Invoke
these functions in main and observe the result