Week2.cpp
Week 2 (structures, character pointers, overloaded functions)
(1) Define a structure to represent a Student. Keep track of name, id and cgpa. Write functions
namely input, display and edit to input details to a Student variable, to display it and to edit it
respectively. Write a main function which defines a student variable. Invoke these functions in
main and observe the result.
(2) Define the same Student structure. Also keep an array to hold 4 Student records. Write
functions namely input, display, arrange and search which inputs records to the array, displays
the array in the order of input, arranges the records in order of id’s and searches for a record
taking id as the key. Provide a suitable menu for this program.
(3) Write a user defined function which determines the length of a given string. The only
parameter to the function is the string. The return value of the function should be the length.
Show the usage of this function in the main.
(4) Write a user defined function which concatenates one string to another. Pass the two strings
as parameters to the function. The return value of the function should be the concatenated string.
Show the usage of this function in the main.
(5) Write a user defined function which compares 2 strings and determines whether the first one
is greater, smaller or equal to the second string. If they are equal return 0, if first string is greater
return 1 and if the first string is lesser return -1. Pass the two strings to be compared as